+++ src/monopoly/IPlayer.java
@@ -11,25 +11,25 @@ public interface IPlayer extends Serializable, Host, GameObject {
     Place getCurrentPlace();
     boolean isReversed();
     void reverse();
-    int getCash();
-    int getDeposit();
-    int getTotalPossessions();
-    void addPossession(Supplier<Integer> possession);
+    double getCash();
+    double getDeposit();
+    double getTotalPossessions();
+    void addPossession(Supplier<Double> possession);
     void addPropertySeller(Consumer1<Consumer0> cb);
 
     void init();
     void giveUp();
-    void changeCash(int amount, String msg);
-    void changeDeposit(int amount, String msg);
+    void changeCash(double amount, String msg);
+    void changeDeposit(double amount, String msg);
     void depositOrWithdraw(Consumer0 cb);
-    void pay(IPlayer receiver, int amount, String msg, Consumer0 cb);
+    void pay(IPlayer receiver, double amount, String msg, Consumer0 cb);
     void startWalking(int steps);
 
     default void startTurn(Consumer0 cb) {
         cb.run();
     }
 
-    default void askHowMuchToDepositOrWithdraw(Consumer1<Integer> cb) {
-        cb.run(0);
+    default void askHowMuchToDepositOrWithdraw(Consumer1<Double> cb) {
+        cb.run(0.0);
     }
 }
