+++ src/monopoly/card/GodOfFortuneCard.java
@@ -11,7 +11,7 @@ public class GodOfFortuneCard extends Card {
     private static final Card instance = new GodOfFortuneCard();
     static {
         Game.putDefaultConfig("god-of-fortune-card-price", 13);
-        Game.putDefaultConfig("god-of-fortune-card-award", 10000);
+        Game.putDefaultConfig("god-of-fortune-card-award", 10000.0);
         Game.putDefaultConfig("god-of-fortune-card-duration", 8);
     }
 
@@ -27,7 +27,7 @@ public class GodOfFortuneCard extends Card {
     @Override
     public void use(Game g, Consumer1<Boolean> cb) {
         IPlayer player = g.getCurrentPlayer();
-        int amount = g.getConfig("god-of-fortune-card-award");
+        double amount = g.getConfig("god-of-fortune-card-award");
         String msg = g.format("blessed_by_god_of_fortune", player.getName(), amount);
         player.changeCash(amount, msg);
         new RentFree(player, g.getConfig("god-of-fortune-card-duration"));
