+++ src/monopoly/card/TaxCard.java
@@ -30,7 +30,7 @@ public class TaxCard extends Card {
                 int reach = g.getConfig("tax-card-reach");
                 if (player != null &&
                         Place.withinReach(current.getCurrentPlace(), player.getCurrentPlace(), reach) >= 0) {
-                    int amount = player.getDeposit() * 3 / 10;
+                    double amount = player.getDeposit() * 3 / 10;
                     String msg = g.format("pay_tax", player.getName(), amount);
                     player.changeDeposit(-amount, msg);
                     cb.run(true);
