+++ src/monopoly/extension/BankSystem.java
@@ -2,6 +2,7 @@ package monopoly.extension;
 
 import monopoly.Game;
 import monopoly.IPlayer;
+import monopoly.Util;
 import monopoly.util.Parasite;
 
 public final class BankSystem {
@@ -16,7 +17,7 @@ public final class BankSystem {
             GameCalendar.onMonth.get(g).addListener(() -> {
                 synchronized (g.lock) {
                     for (IPlayer player : g.getPlayers()) {
-                        int amount = player.getDeposit() / 10;
+                        double amount = player.getDeposit() / 10;
                         String msg = g.format("give_interest", player.getName(), amount);
                         player.changeDeposit(amount, msg);
                     }
